WIP zoekt-distributed refactor 20250901 0932
